# Support nullable group by