--- a/common/datavalues/src/types/deserializations/date.rs
+++ b/common/datavalues/src/types/deserializations/date.rs
@@ -32,7 +32,7 @@ where
     T: PrimitiveType,
     T: Unmarshal<T> + StatBuffer + FromLexical,
 {
-    fn de(&mut self, reader: &mut &[u8]) -> Result<()> {
+    fn de_binary(&mut self, reader: &mut &[u8]) -> Result<()> {
         let value: T = reader.read_scalar()?;
         self.builder.append_value(value);
         Ok(())
@@ -42,7 +42,7 @@ where
         self.builder.append_value(T::default());
     }
 
-    fn de_batch(&mut self, reader: &[u8], step: usize, rows: usize) -> Result<()> {
+    fn de_fixed_binary_batch(&mut self, reader: &[u8], step: usize, rows: usize) -> Result<()> {
         for row in 0..rows {
             let mut reader = &reader[step * row..];
             let value: T = reader.read_scalar()?;