--- a/common/datavalues/src/types/deserializations/date_time.rs
+++ b/common/datavalues/src/types/deserializations/date_time.rs
@@ -37,7 +37,7 @@ where
         common_clickhouse_srv::types::Simple,
     >,
 {
-    fn de(&mut self, reader: &mut &[u8]) -> Result<()> {
+    fn de_binary(&mut self, reader: &mut &[u8]) -> Result<()> {
         let value: T = reader.read_scalar()?;
         self.builder.append_value(value);
         Ok(())
@@ -47,7 +47,7 @@ where
         self.builder.append_value(T::default());
     }
 
-    fn de_batch(&mut self, reader: &[u8], step: usize, rows: usize) -> Result<()> {
+    fn de_fixed_binary_batch(&mut self, reader: &[u8], step: usize, rows: usize) -> Result<()> {
         for row in 0..rows {
             let mut reader = &reader[step * row..];
             let value: T = reader.read_scalar()?;