… query/src/pipelines/transforms/group_by/aggregator_groups_builder.rs
… query/src/pipelines/transforms/transform_group_by_final.rs
… tests/suites/0_stateless/03_dml/03_0003_select_group_by.result
… tests/suites/0_stateless/03_dml/03_0003_select_group_by.sql