I hereby agree to the terms of the CLA available at: https://databend.rs/dev/policies/cla/

## Summary

Summary about this PR

## Changelog

 
- Improvement
 

## Related Issues

Continue: #4158
Fixes #4079 

## Test Plan

Unit Tests

Stateless Tests