@sundy-li 2022-03-06
Ontime datasets benchmark (the second query with metacache)：

Before:

```
MySQL [(none)]> SELECT DayOfWeek, count(*) AS c FROM ontime GROUP BY DayOfWeek ORDER BY c DESC;
+-----------+----------+
| DayOfWeek | c        |
+-----------+----------+
|         3 | 23513895 |
|         2 | 23384295 |
|         4 | 23113350 |
|         7 | 21270195 |
|         1 | 20861550 |
|         6 | 19360215 |
|         5 | 18896085 |
+-----------+----------+
7 rows in set (2.578 sec)
```

After:

```
MySQL [(none)]> SELECT DayOfWeek, count(*) AS c FROM ontime GROUP BY DayOfWeek ORDER BY c DESC;
+-----------+----------+
| DayOfWeek | c        |
+-----------+----------+
|         3 | 23513895 |
|         2 | 23384295 |
|         4 | 23113350 |
|         7 | 21270195 |
|         1 | 20861550 |
|         6 | 19360215 |
|         5 | 18896085 |
+-----------+----------+
7 rows in set (0.896 sec)
```