@BohuTANG 2022-03-06
Also cc @wubx @dantengsky for the performance improvement.
 
It's time to change the subject/ontime table schema:
https://github.com/datafuselabs/databend/discussions/4342